Code Refactoring and Modularization started.
